+++ src/components/PupilViews/PupilProgrammeListing/PupilProgrammeListing.view.tsx
@@ -13,29 +13,27 @@ import {
   oakDefaultTheme,
 } from "@oaknational/oak-components";
 
-import { getExamboardData } from "./getExamboardData";
-
 import { PupilProgrammeListingData } from "@/node-lib/curriculum-api-2023/queries/pupilProgrammeListing/pupilProgrammeListing.schema";
 import {
-  BrowseExamboardSelector,
-  ExamboardData,
-} from "@/components/PupilComponents/BrowseExamboardSelector";
-import {
-  BrowseTierSelector,
-  TierData,
-} from "@/components/PupilComponents/BrowseTierSelector";
+  BrowseFactorSelector,
+  Factors,
+} from "@/components/PupilComponents/BrowseFactorSelector";
 import { resolveOakHref } from "@/common-lib/urls";
 import AppLayout from "@/components/SharedComponents/AppLayout";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import SignpostTeachersInlineBanner from "@/components/PupilComponents/SignpostTeachersInlineBanner/SignpostTeachersInlineBanner";
+import { FactorData } from "@/pages-helpers/pupil/options-pages/getAvailableProgrammeFactor";
+import { getFactorDataFromSlug } from "@/pages-helpers/pupil/options-pages/getFactorDataFromSlug";
 
 export type PupilViewsProgrammeListingProps = {
   programmes: PupilProgrammeListingData[];
   baseSlug: string;
   yearSlug: PupilProgrammeListingData["yearSlug"];
-  examboardSlug?: ExamboardData["examboardSlug"];
-  examboards: ExamboardData[];
-  tiers: TierData[];
+  examboardSlug?: PupilProgrammeListingData["programmeFields"]["examboardSlug"];
+  examboards: FactorData[];
+  tiers: FactorData[];
+  pathwaySlug?: PupilProgrammeListingData["programmeFields"]["pathwaySlug"];
+  pathways: FactorData[];
 };
 
 export const PupilViewsProgrammeListing = ({
@@ -45,12 +43,53 @@ export const PupilViewsProgrammeListing = ({
   examboardSlug,
   examboards,
   tiers,
+  pathwaySlug,
+  pathways,
 }: PupilViewsProgrammeListingProps) => {
-  const [chosenExamboard, setChosenExamboard] = useState<ExamboardData | null>(
-    examboardSlug && examboards.length >= 1
-      ? getExamboardData({ examboardSlug, availableExamboards: examboards })
-      : null,
+  const orderedFactors: ("pathway" | "examboard" | "tier")[] = [
+    "pathway",
+    "examboard",
+    "tier",
+  ];
+
+  const [chosenFactors, setChosenFactors] = useState<Factors>({
+    pathway:
+      pathwaySlug && pathways.length >= 1
+        ? getFactorDataFromSlug({
+            factorSlug: pathwaySlug,
+            availableFactors: pathways,
+          })
+        : null,
+    examboard:
+      examboardSlug && examboards.length >= 1
+        ? getFactorDataFromSlug({
+            factorSlug: examboardSlug,
+            availableFactors: examboards,
+          })
+        : null,
+    tier: null,
+  });
+
+  const allFactors = {
+    pathway: pathways,
+    examboard: examboards,
+    tier: tiers,
+  };
+
+  const availableFactors = orderedFactors.filter(
+    (f) =>
+      (f === "pathway" && pathways.length > 1) ||
+      (f === "tier" && tiers.length > 1) ||
+      (f === "examboard" && examboards.length > 1),
   );
+
+  const currentFactor =
+    availableFactors.filter(
+      (f) =>
+        (f === "pathway" && chosenFactors.pathway === null) ||
+        (f === "examboard" && chosenFactors.examboard === null) ||
+        (f === "tier" && chosenFactors.tier === null),
+    )[0] || "none";
 
   if (!programmes[0]) {
     throw new Error("No programme data available");
@@ -67,80 +106,96 @@ export const PupilViewsProgrammeListing = ({
   const yearDescriptions = programmes[0]?.programmeFields.yearDescription;
 
   const topNavSlot = () => {
-    if (chosenExamboard !== null && tiers.length > 1) {
-      return (
-        <OakTertiaryButton
-          iconName="arrow-left"
-          onClick={() => setChosenExamboard(null)}
-        >
-          Change examboard
-        </OakTertiaryButton>
-      );
-    } else {
-      return (
-        <OakTertiaryButton
-          iconName="arrow-left"
-          element="a"
-          href={resolveOakHref({
-            page: "pupil-subject-index",
-            yearSlug,
-          })}
-        >
-          Change subject
-        </OakTertiaryButton>
-      );
+    const currentIndex =
+      currentFactor !== "none" ? availableFactors.indexOf(currentFactor) : -1;
+    const prevIndex = currentIndex - 1;
+    const option = availableFactors[prevIndex] || "none";
+
+    switch (option) {
+      case "examboard":
+      case "tier":
+      case "pathway":
+        return (
+          <OakTertiaryButton
+            iconName="arrow-left"
+            onClick={() =>
+              setChosenFactors({ ...chosenFactors, [option]: null })
+            }
+          >
+            Change {option}
+          </OakTertiaryButton>
+        );
+      default:
+        return (
+          <OakTertiaryButton
+            iconName="arrow-left"
+            element="a"
+            href={resolveOakHref({
+              page: "pupil-subject-index",
+              yearSlug,
+            })}
+          >
+            Change subject
+          </OakTertiaryButton>
+        );
     }
   };
 
   const breadcrumbs: string[] = [yearDescriptions];
-  if (chosenExamboard?.examboard && tiers.length > 1) {
-    breadcrumbs.push(chosenExamboard.examboard);
-  }
+
+  orderedFactors.forEach((f) => {
+    if (chosenFactors[f]?.factor) {
+      breadcrumbs.push(chosenFactors[f].factor);
+    }
+  });
 
   const BrowseOptions = () => {
-    if (
-      (chosenExamboard !== null && tiers.length > 1) ||
-      (examboards.length <= 1 && tiers.length >= 1)
-    ) {
-      return (
-        <BrowseTierSelector
-          tiers={tiers}
-          baseSlug={baseSlug}
-          examboardSlug={
-            chosenExamboard ? chosenExamboard.examboardSlug : undefined
-          }
-          phaseSlug={phaseSlug}
-        />
-      );
-    } else if (examboards.length > 1 && chosenExamboard === null) {
-      return (
-        <BrowseExamboardSelector
-          examboards={examboards}
-          baseSlug={baseSlug}
-          onClick={
-            tiers.length > 1
-              ? (examboard) => setChosenExamboard(examboard)
-              : undefined
-          }
-          phaseSlug={phaseSlug}
-        />
-      );
-    } else {
+    if (currentFactor === "none") {
       return <OakBox>No programme factors to be selected</OakBox>;
     }
+
+    if (currentFactor) {
+      return (
+        <BrowseFactorSelector
+          factorType={currentFactor}
+          factors={allFactors[currentFactor]}
+          baseSlug={baseSlug}
+          chosenFactors={chosenFactors}
+          programmes={programmes}
+          onClick={(f) =>
+            setChosenFactors({ ...chosenFactors, [currentFactor]: f })
+          }
+          phaseSlug={phaseSlug}
+        />
+      );
+    }
+
+    return <OakBox>No programme factors to be selected</OakBox>;
   };
 
   const optionTitles = (): { hint: string; title: string } => {
-    if (examboards.length > 1 && chosenExamboard === null) {
-      return {
-        hint: "An exam board sets and assesses what you learn in your GCSE course. If you can't find your exam board here or are not sure which exam board you are studying, ask your teacher.",
-        title: "Choose an exam board",
-      };
+    switch (currentFactor) {
+      case "pathway":
+        return {
+          hint: "A pathway is either examined or non-examined",
+          title: "Choose a pathway",
+        };
+      case "examboard":
+        return {
+          hint: "An exam board sets and assesses what you learn in your GCSE course. If you can't find your exam board here or are not sure which exam board you are studying, ask your teacher.",
+          title: "Choose an exam board",
+        };
+      case "tier":
+        return {
+          hint: "Some GCSE subjects have a foundation and higher tier. Different content may be covered in each tier so it is important to check with your teacher if you are unsure which tier you are doing.",
+          title: "Choose a tier",
+        };
+      default:
+        return {
+          hint: "",
+          title: "No programme factors to be selected",
+        };
     }
-    return {
-      hint: "Some GCSE subjects have a foundation and higher tier. Different content may be covered in each tier so it is important to check with your teacher if you are unsure which tier you are doing.",
-      title: "Choose a tier",
-    };
   };
 
   const iconSlug = `subject-${subjectSlug}`;