+++ src/node-lib/curriculum-api-2023/queries/curriculumUnits/curriculumUnits.query.ts
@@ -25,6 +25,7 @@ const curriculumUnitsQuery =
           ],
         },
         { phase_slug: { _eq: phaseSlug } },
+        { state: { _eq: "published" } },
       ],
     };
 
@@ -37,7 +38,17 @@ const curriculumUnitsQuery =
       ? {
           _or: [
             { examboard_slug: { _eq: examboardSlug } },
-            { examboard_slug: { _is_null: true } },
+            {
+              _and: [
+                { examboard_slug: { _is_null: true } },
+                {
+                  _or: [
+                    { pathway_slug: { _neq: "core" } },
+                    { pathway_slug: { _is_null: true } },
+                  ],
+                },
+              ],
+            },
           ],
         }
       : { examboard_slug: { _is_null: true } };